+++ CHANGELOG.md
@@ -6,7 +6,14 @@
 - group scores?
 - Exception in thread "Thread-9"
 Exception: java.lang.NoClassDefFoundError thrown from the UncaughtExceptionHandler in thread "Thread-9"
-- [BUG] user name substring match.
+- BUG: user names with their substrings.
+
+## 0.4.2 - 2021-12-13
+- removed development line in page.clj.
+
+## 0.4.1 - 2021-12-13
+- members must be three.
+- remove libraries that are not used.
 
 ## 0.4.0 - 2021-12-13
 - forbidden duplicate same user name in a group.